run migration as task
